python-service/requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
bitsandbytes>=0.41.0
accelerate>=0.24.0

# Vector Search (FAISS for ultra-fast similarity search)
# FAISS CPU provides significant speedup over PyTorch for large embedding sets
# Note: GPU acceleration is handled via PyTorch tensors, FAISS runs on CPU
faiss-cpu>=1.7.4

# PDF Processing
pymupdf>=1.23.0

# Text Processing
nltk>=3.8.0
rank-bm25>=0.2.2

# Code Processing (Phase 2)
tree-sitter>=0.20.4

# Image Processing (Phase 2)
Pillow>=10.0.0
pytesseract>=0.3.10

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# Utilities
numpy>=1.24.0
python-magic>=0.4.27

# API Validation
pydantic>=2.0.0

